refactor: move Theme back to view, modify imports
