+++ app/src/main/kotlin/view/Theme.kt
@@ -1,9 +1,7 @@
-package view.theme
+package view
 
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import view.DarkColorPalette
-import view.LightColorPalette
 
 @Composable
 fun Material3AppTheme(theme: Theme, content: @Composable () -> Unit) {
