+++ app/src/main/kotlin/view/mainScreen/ThemeSwitcher.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import view.theme.Theme
+import view.Theme
 
 @Composable
 fun ThemeSwitcher(
